Update Filestore docs to describe new runner file naming
There is a small difference in the way the Runner creates the
encrypted_user_id_and_token.

It still uses the service secret as the key when encrypting, but the
secret itself is not part of what is being encrypted.

It is also does not use the service slug when creating the final hashed
key.

This is effectively just the name of the file that sits on S3. The file
payload is encrypted in the same way before upload. It is also encyrpted
at rest in the bucket.

The user_id and user_token as also remain encrypted